move `SurveyResultBadge` `@tamanu/ui-components`
